# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remedy Scribe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/encounters/{encounter_id}/upload/init": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Init Upload",
        "operationId": "init_upload_api_v1_encounters__encounter_id__upload_init_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/{encounter_id}/upload/parts/{part_number}": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Part Upload Url",
        "operationId": "get_part_upload_url_api_v1_encounters__encounter_id__upload_parts__part_number__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          },
          {
            "name": "part_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "title": "Part Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartUploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/{encounter_id}/upload/parts": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "List Upload Parts",
        "description": "What a resuming client diffs against \u2014 see storage.list_uploaded_parts:\nS3 is the source of truth for \"what's already landed,\" not a\nPostgres table this app would have to keep in sync with it.",
        "operationId": "list_upload_parts_api_v1_encounters__encounter_id__upload_parts_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPartsStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/{encounter_id}/upload/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload",
        "description": "Finalizes the S3 object, then runs the same gate + bookkeeping the\nold confirm_upload did: consent check first (before touching S3 \u2014\nan invalid-consent encounter's multipart upload is left incomplete\nfor the lifecycle rule to reap, not finalized), then retention clock,\npipeline_status, and kicking the pipeline.",
        "operationId": "complete_upload_api_v1_encounters__encounter_id__upload_complete_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EncounterOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "pipeline_status": {
            "$ref": "#/components/schemas/EncounterPipelineStatus"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "last_pipeline_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Pipeline Error"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "audio_retention_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Retention Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "pipeline_status",
          "retry_count",
          "last_pipeline_error",
          "audio_retention_expires_at",
          "created_at"
        ],
        "title": "EncounterOut"
      },
      "EncounterPipelineStatus": {
        "type": "string",
        "enum": [
          "recording",
          "uploaded",
          "transcribed",
          "note_generated",
          "blocked_no_consent",
          "transcription_failed",
          "generation_failed"
        ],
        "title": "EncounterPipelineStatus",
        "description": "Phase 0.4: was a free-form String(32) that the codebase wrote at\nleast five different values into across two files (encounters.py,\ntasks/pipeline.py) with nothing checking any of them matched. Now a\nproper enum, the way Note.status already was in name \u2014 see the\n`create_constraint=True` below for why \"already was\" turned out to\nbe only half true."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartUploadUrlResponse": {
        "properties": {
          "part_number": {
            "type": "integer",
            "title": "Part Number"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          }
        },
        "type": "object",
        "required": [
          "part_number",
          "url",
          "expires_in_seconds"
        ],
        "title": "PartUploadUrlResponse"
      },
...
      "UploadInitRequest": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          }
        },
        "type": "object",
        "title": "UploadInitRequest",
        "description": "`content_type` is optional and not enforced against an allowlist\n\u2014 Phase 2.2 hasn't picked a recording codec yet, so this schema\ndoesn't get to invent one. It's used only to pick a file extension\nand set the S3 object's Content-Type metadata."
      },
      "UploadInitResponse": {
        "properties": {
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "min_part_size_bytes": {
            "type": "integer",
            "title": "Min Part Size Bytes"
          },
          "max_part_number": {
            "type": "integer",
            "title": "Max Part Number"
          }
        },
        "type": "object",
        "required": [
          "object_key",
          "upload_id",
          "min_part_size_bytes",
          "max_part_number"
        ],
        "title": "UploadInitResponse"
      },
      "UploadPartsStatusResponse": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/UploadedPart"
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "type": "object",
        "required": [
          "parts"
        ],
        "title": "UploadPartsStatusResponse",
        "description": "What a resuming client diffs its own local chunk manifest\nagainst to know which parts it can skip re-uploading."
      },
      "UploadedPart": {
        "properties": {
          "part_number": {
            "type": "integer",
            "title": "Part Number"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "etag": {
            "type": "string",
            "title": "Etag"
          }
        },
        "type": "object",
        "required": [
          "part_number",
          "size_bytes",
          "etag"
        ],
        "title": "UploadedPart"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}